chore(common): Rename http-utils.ts to fetch-json.ts
