+++ packages/common/src/index.ts
@@ -3,7 +3,7 @@ export * from './ceramic-api.js'
 export * from './context.js'
 export * from './streamopts.js'
 export * from './stream.js'
-export * from './utils/http-utils.js'
+export * from './utils/fetch-json.js'
 export * from './utils/stream-utils.js'
 export * from './utils/test-utils.js'
 export * from './utils/accountid-utils.js'
